pyproject.toml
[tool.poetry]
name = "habit-reminder"
version = "0.1.0"
description = ""
authors = ["leonid Petrov <leonidpetrov637@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
isort = "^5.13.2"
flake8 = "^7.1.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
